feat: sql main with search and origin/destination
