+++ app/[slug]/page.tsx
@@ -1,28 +1,13 @@
-import mongoose from "mongoose";
+import { QueryResultRow, sql } from "@vercel/postgres";
+import { unstable_noStore as noStore } from "next/cache";
 import Grid from "./grid";
-
-async function getDb(col: string) {
-  try {
-    const mongo = await mongoose.connect(process.env.DB, { dbName: "Tickets" });
-    const db = await mongo.connection.db;
-    const collection = db.collection(col);
-    const data = await collection.find().sort({ seatNumber: 1 }).toArray();
-    const totalDocs = await collection.countDocuments();
-    await mongoose.connection.close();
-    return { data};
-  } catch (e) {
-    console.log(e);
-    throw new Error("Connection failed");
-  }
-}
-
-export default async function Home() {
-
-
-  const { data} = await getDb("START");
-  const newChunkedArray = []
-  for (let i = 0; i < data.length; i += 6) {
-    const chunk = data.slice(i, i + 6);
+export default async function Home({ params }: { params: { slug: string } }) {
+  noStore();
+  const { rows } =
+    await sql`SELECT * from seats WHERE ticket_id = ${params.slug.toUpperCase()}`;
+  const newChunkedArray: QueryResultRow | number = [];
+  for (let i = 0; i < rows.length; i += 6) {
+    const chunk = rows.slice(i, i + 6);
     newChunkedArray.push(
       ...chunk.slice(0, 3),
       Math.floor(i / 6) + 1,
@@ -30,10 +15,13 @@ export default async function Home() {
     );
   }
   return (
-    <main className="flex text-black min-h-screen flex-col items-center  px-3">
-      <h1 className="pb-12 text-4xl">Select your seat</h1>
-      
-      <Grid data={JSON.parse(JSON.stringify(newChunkedArray))}></Grid>
+    <main className="flex text-black  min-h-screen flex-col items-center py-3 px-3">
+      {/* {rows.map(flight => (
+      <div key={flight.id}>
+        {flight.seat_number} 
+      </div>
+    ))} */}
+      <Grid data={newChunkedArray as Seat[]}></Grid>
     </main>
   );
 }
